Move interpreter to capture module
